Adapt 745  documentation and organization (#59)
* ADAPT-745: Added documentation on component organization.

* ADAPT-745: Reorganized story full view.

* ADAPT-745: Cherry picked and added comments again.

* ADAPT-745: Added more comments for partial files.

* ADAPT-745: Moved components into new folders media and entryPickers.

* ADAPT-745: Reorganizing and merging conflicts to update story card view.

* ADAPT-745: Condensed comments.

* ADAPT-745: Organizing mega menu components into a folder, updated references.

* ADAPT-745: Updated components.js file and organized Content Menu components into a sub folder.

* ADAPT-745: Moved over date and image processing info from story to storyFull component.

* ADAPT-745: Updated required import lines for story components.

* ADAPT-745: Added story subfolder.
