+++ src/components/components.js
@@ -22,17 +22,17 @@ import OodPoster from './composite/oodPoster'
 import OodQuoteSlider from './composite/oodQuoteSlider'
 // Navigation Components
 import NavItem from './navigation/navItem'
-import ContentNavItem from './navigation/contentNavItem'
-import ContentMenuParentItems from './navigation/contentMenuParentItem'
-import ContentNestedMenu from './navigation/contentNestedMenu'
-import OodContentMenu from './navigation/oodContentMenu'
-import ContentMenuPicker from './navigation/contentMenuPicker'
+import ContentNavItem from './navigation/contentMenu/contentNavItem'
+import ContentMenuParentItems from './navigation/contentMenu/contentMenuParentItem'
+import ContentNestedMenu from './navigation/contentMenu/contentNestedMenu'
+import OodContentMenu from './navigation/contentMenu/oodContentMenu'
+import ContentMenuPicker from './navigation/contentMenu/contentMenuPicker'
 import OodSubMenu from './navigation/oodSubMenu'
-import OodMegaMenu from './navigation/oodMegaMenu'
-import OodMegaMenuSection from './navigation/oodMegaMenuSection'
-import OodMegaMenuCard from './navigation/oodMegaMenuCard'
-import OodMegaMenuNavItem from './navigation/oodMegaMenuNavItem'
-import OodMegaMenuLinkGroup from './navigation/oodMegaMenuLinkGroup'
+import OodMegaMenu from './navigation/megaMenu/oodMegaMenu'
+import OodMegaMenuSection from './navigation/megaMenu/oodMegaMenuSection'
+import OodMegaMenuCard from './navigation/megaMenu/oodMegaMenuCard'
+import OodMegaMenuNavItem from './navigation/megaMenu/oodMegaMenuNavItem'
+import OodMegaMenuLinkGroup from './navigation/megaMenu/oodMegaMenuLinkGroup'
 // Layout Components
 import ColumnGrid from './layout/columnGrid'
 import CuratedStories from './layout/curatedStories'
@@ -45,14 +45,14 @@ import SingleColumnContent from './layout/singleColumnContent'
 // Page Type Components
 import OodInteriorPage from './page-types/oodInteriorPage'
 import OodLandingPage from './page-types/oodLandingPage'
-import OodStory from './page-types/oodStory'
+import OodStory from './page-types/story/oodStory'
 import OodSupportPage from './page-types/oodSupportPage'
 // Simple Components
-import AspectRatioImage from './simple/aspectRatioImage'
+import AspectRatioImage from './media/aspectRatioImage'
 import CtaLink from './simple/ctaLink'
 import EmbedScript from './simple/embedScript'
-import StoryImage from './simple/storyImage'
-import StoryPicker from './simple/storyPicker'
+import StoryImage from './media/storyImage'
+import StoryPicker from './page-types/story/storyPicker'
 // Composite Components
 import Accordion from './composite/accordion'
 import AccordionItem from './composite/accordionItem'
