+++ src/components/media/aspectRatioImage.js
