+++ src/components/media/storyImage.js
