+++ src/components/partials/belowContent.js
@@ -2,6 +2,8 @@ import SbEditable from "storyblok-react";
 import React from "react";
 import Components from "../components"
 
+/* The BelowContent component is referenced by the Interior Page and Support page types. */
+
 const BelowContent = (props) => (
   <SbEditable content={props.blok}>
     {(props.blok.belowContent != null && Object.keys(props.blok.belowContent).length > 0) && (
