+++ src/components/partials/footer.js
@@ -2,6 +2,11 @@ import React from "react";
 import SbEditable from "storyblok-react";
 import Components from "../components";
 
+/*
+** The footer component is referenced and used in all page-type components.
+** It incorporates the local footer and global footer, based on page settings.
+*/
+
 const Footer = (props) => (
   <SbEditable content={props.blok}>
     <footer>
