+++ src/components/partials/headerMinimal.js
@@ -1,6 +1,8 @@
 import SbEditable from "storyblok-react";
 import React from "react";
 
+/* The Header Minimal component is referenced by the Interior Page type. */
+
 const HeaderMinimal = (props) => (
   <SbEditable content={props.blok}>
     <header className={`ood-interior-page__header ood-interior-page__header--minimal su-bg-fog-light su-py-7`}>
