+++ src/components/partials/headerNoImage.js
@@ -2,6 +2,8 @@ import React from "react";
 import SbEditable from "storyblok-react";
 import RichTextField from "../../utilities/richTextField";
 
+/* The Header No Image component is referenced by the Interior Page and Support page types. */
+
 const HeaderNoImage = (props) => (
   <SbEditable content={props.blok}>
     <header className={`ood-interior-page__header ood-interior-page__header`}>
