+++ src/components/partials/htmlHead.js
@@ -1,5 +1,11 @@
 import React from 'react'
 
+/*
+** The HtmlHead component is included into the html.js file, to always be loaded.
+* The purpose of this file is to provide a default Title value,
+* ensure that favicons are provided from Stanford's media source, particularly for browser tab and phone home screens.
+*/
+
 const HtmlHead = (props) => {
   return (
     <>
